% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{The input data for example}
\format{
A matrix with 10 rows and 5 columns.
}
\usage{
example_data
}
\description{
In this matrix, columns represent the results of different clustering results and rows are samples.
}
\keyword{datasets}
