% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralClusteringAffinity.R
\name{spectralClusteringAffinity}
\alias{spectralClusteringAffinity}
\title{Perform spectral clustering algorithms for an affinity matrix, using SNFtool::spectralClustering.}
\usage{
spectralClusteringAffinity(affi_matrix, k, type = 3)
}
\arguments{
\item{affi_matrix}{A numerical similarity or affinity matrix.}

\item{k}{A number value of clusters}

\item{type}{The variants of spectral clustering to use. See \code{SNFtool::spectralClustering}}
}
\value{
A vector consisting of cluster labels of each sample.
}
\description{
Perform spectral clustering algorithms for an affinity matrix, using SNFtool::spectralClustering.
}
